pyproject.toml
[project]
name = "testgenflow"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "requests>=2.32",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"
norecursedirs = ["artifacts", ".artifacts", ".git", ".venv"]

[tool.black]
line-length = 100

[tool.mypy]
python_version = "3.10"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
exclude = ["artifacts"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
